Move test framework plugins out of lib/
