+++ index.js
@@ -7,7 +7,7 @@ module.exports = {
   setFirebaseData: helpers.setFirebaseData,
   setFirebaseRules: helpers.setFirebaseRules,
   setDebug: helpers.setDebug,
-  users: helpers.userDefinitions,
+  users: helpers.users,
   utils: helpers,
   chai: require('./lib/chai'),
   jasmine: require('./lib/jasmine'),
