+++ lib/test-cmd.js
