+++ test/spec/.eslintrc.yml
@@ -1,4 +1,4 @@
-extends: '../../../.eslintrc.yml'
+extends: '../../.eslintrc.yml'
 env:
   mocha: true
 globals:
