+++ test/spec/plugins/chai.js
@@ -2,12 +2,12 @@
 'use strict';
 
 var chai = require('chai'),
-  plugin = require('../../../lib/chai');
+  targaryen = require('../../../plugins/chai');
 
 describe('Chai plugin', function() {
 
   before(function() {
-    chai.use(plugin);
+    chai.use(targaryen);
   });
 
   describe('if unconfigured', function() {
@@ -26,7 +26,7 @@ describe('Chai plugin', function() {
 
     it('sets the underlying Firebase database to the supplied data', function() {
 
-      plugin.setFirebaseData({
+      targaryen.setFirebaseData({
         users: {
           'password:500f6e96-92c6-4f60-ad5d-207253aee4d3': {
             name: 'Sherlock Holmes'
@@ -45,7 +45,7 @@ describe('Chai plugin', function() {
 
     it('sets the underlying Firebase database to the supplied data', function() {
 
-      plugin.setFirebaseRules({
+      targaryen.setFirebaseRules({
         rules: {
           users: {
             '$user': {
@@ -77,7 +77,7 @@ describe('Chai plugin', function() {
 
     beforeEach(function() {
 
-      plugin.setFirebaseData({
+      targaryen.setFirebaseData({
         users: {
           'password:500f6e96-92c6-4f60-ad5d-207253aee4d3': {
             name: 'Sherlock Holmes'
@@ -88,7 +88,7 @@ describe('Chai plugin', function() {
         }
       });
 
-      plugin.setFirebaseRules({
+      targaryen.setFirebaseRules({
         rules: {
           users: {
             '$user': {
@@ -148,7 +148,7 @@ describe('Chai plugin', function() {
     });
 
     it('should permit patch tests', function() {
-      plugin.setFirebaseData({
+      targaryen.setFirebaseData({
         users: {
           'password:500f6e96-92c6-4f60-ad5d-207253aee4d3': {
             name: 'Sherlock Holmes'
